requirements_dev.txt
apache-airflow==2.3.3
pre-commit==2.16.0
pre-commit-hooks==4.0.1
flit==3.5.1
flit-core==3.5.1
black==20.8b1
coverage==6.2
flake8==4.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pylint==3.0.0a1
pytest==6.2.5
tox==3.24.4
Sphinx==4.3.1
pip-licenses==3.5.3
sphinx_rtd_theme
sphinx-bootstrap-theme==0.8.0
sphinxcontrib-plantuml
pytest-html==3.1.1
